Updated adequate reaction to misgendering
Asking someone to apologize for their behavior may make the situation worse: they may not mean the apology, which would show, or not really understand the issue, which would also make the apologize vain. Also and most importantly, the impacted person's well-being should be priority number 1 and they may not want to interact with the actor again.